# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XConnect API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/inventory/wallet": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get technician's current inventory wallet",
                "operationId": "df046b91e6464caf78d681d4fb1de403",
                "responses": {
                    "200": {
                        "description": "Wallet items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Inventory item ID",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Item name",
                                                "type": "string"
                                            },
                                            "sku": {
                                                "description": "Stock keeping unit",
                                                "type": "string"
                                            },
                                            "type": {
                                                "description": "Item type (connector, cable, etc.)",
                                                "type": "string"
                                            },
                                            "quantity": {
                                                "description": "Current quantity in wallet",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory/requests": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Submit an inventory replenishment request",
                "operationId": "daad494b9e0719464bc2ab218cd5917a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pickup_date",
                                    "pickup_location",
                                    "items"
                                ],
                                "properties": {
                                    "pickup_date": {
                                        "description": "Date to pick up inventory (today or future)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-22"
                                    },
                                    "pickup_location": {
                                        "description": "Warehouse location for pickup",
                                        "type": "string",
                                        "enum": [
                                            "Daytona Beach",
                                            "Melbourne",
                                            "Fort Pierce"
                                        ],
                                        "example": "Daytona Beach"
                                    },
                                    "notes": {
                                        "description": "Additional notes for the request",
                                        "type": "string",
                                        "example": "Need ASAP for new installs",
                                        "nullable": true
                                    },
                                    "items": {
                                        "description": "Inventory items requested",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "description": "Inventory item ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "description": "Quantity requested",
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request submitted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Use Sanctum bearer token from /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}